Extracted RestDataBuilder
